refactor custom errors to new errors module
